backend_app/requirements.txt
aiogram==3.0.0b4
aiohttp_cors==0.7.0
alembic==1.8.1
asyncpg==0.26.0
bcrypt==4.0.0
email-validator==1.3.0
fastapi==0.85.0
fastapi-pagination==0.11.0
jwt==1.3.1
multidict==6.0.2
passlib==1.7.4
psycopg2-binary==2.9.5
pydantic==1.10.2
python-jose==3.3.0
python-multipart==0.0.5
shortuuid==1.0.9
SQLAlchemy==1.4.41
typing_extensions==4.3.0
uuid==1.30
uvicorn==0.18.3
websockets==10.3
requests==2.28.1
pytz